Move Workshops to component
